feat: Implementando controle de fluxo (Marca)

+++ app/Http/Controllers/MarcaController.php
@@ -34,14 +34,19 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca->toArray();
+
+        return (!$marca) ? ['error' => 'Recurso solicitado indisponível.'] : $marca->toArray();
     }
 
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+
+        if(!$marca) return ['error' => 'Impossível realizar esta atualização. O recurso solicitado não existe.'];
+
         $marca->nome = $request->get('nome');
         $marca->imagem = $request->get('imagem');
+
         $marca->save();
 
         return $marca->toArray();
@@ -50,6 +55,9 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if(!$marca) return ['error' => 'Impossível realizar esta remoção. O recurso solicitado não existe.'];
+
         $marca->delete();
 
         return ["message" => 'Marca "' . $marca->nome . '" foi removida com sucesso!'];
